@@ -11,6 +11,7 @@
     },
     "devDependencies": {
         "@popperjs/core": "^2.10.2",
+        "@vue/compiler-sfc": "^3.2.45",
         "axios": "^0.25",
         "bootstrap": "^5.1.3",
         "laravel-mix": "^6.0.6",
@@ -19,13 +20,12 @@
         "resolve-url-loader": "^3.1.2",
         "sass": "^1.32.11",
         "sass-loader": "^11.0.1",
-        "vue-loader": "^15.9.8",
-        "vue-template-compiler": "^2.1.8"
+        "vue-loader": "^17.0.1"
     },
     "dependencies": {
         "bootstrap-vue": "^2.22.0",
-        "vue": "^2.0.8",
+        "vue": "^3.2.36",
         "vue-axios": "^3.4.1",
-        "vue-router": "^2.2.1"
+        "vue-router": "^4.0.13"
     }
 }
