@@ -10,7 +10,7 @@
 
 @include('footer')
 
-<script src="./js/app.js"></script>
+<script src="/js/app.js"></script>
 <script src="https://cdn.jsdelivr.net/npm/bootstrap@5.1.3/dist/js/bootstrap.bundle.min.js" crossorigin="anonymous"></script>
 </body>
 
